==> notebooks/00_data_split.ipynb <==
import pandas as pd 

df = pd.read_csv(r"E:\ml_projects\e2e_house_price_estimator\data\raw\HouseTS.csv")
df.head(1)

df.shape

df.columns

df.info()

df.date = pd.to_datetime(df.date) 


df.date.dt.year.value_counts().sort_index()

df = df.sort_values("date")
df.head(1)

# Define cutoff dates
cutoff_date_eval = "2020-01-01"   # validation starts
cutoff_date_holdout = "2022-01-01"  # holdout starts

# Train: before 2020
train_df = df[df["date"] < cutoff_date_eval]

# Validation: 2020–2021
eval_df = df[(df["date"] >= cutoff_date_eval) & (df["date"] < cutoff_date_holdout)]

# Holdout: 2022–2023
holdout_df = df[df["date"] >= cutoff_date_holdout]

print("Train shape:", train_df.shape)
print("Eval shape:", eval_df.shape)
print("Holdout shape:", holdout_df.shape)

train_df.to_csv(r"../data/raw/train.csv", index=False)
eval_df.to_csv(r"../data/raw/eval.csv", index=False)
holdout_df.to_csv(r"../data/raw/holdout.csv", index=False)
