==> notebooks/04_baseline.ipynb <==
## Imports

import pandas as pd 
import numpy as np 
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from sklearn.dummy import DummyRegressor

## Load Data

train_df = pd.read_csv(r"E:\ml_projects\e2e_house_price_estimator\data\interim\eval_1.csv")
eval_df = pd.read_csv(r"E:\ml_projects\e2e_house_price_estimator\data\interim\eval_1.csv")

## Define Target and Features

# ============================================
# 3. Define target & features
# ============================================
target = "price"
X_train = train_df.drop(columns=[target])
y_train = train_df[target]

X_val = eval_df.drop(columns=[target])
y_val = eval_df[target]

## Train Baseline Model

baseline = DummyRegressor(strategy="median")
baseline.fit(X_train, y_train)
y_pred = baseline.predict(X_val)


## Model Evaluation

print("Baseline R2 score:", r2_score(y_val, y_pred))
print("Baseline MAE:", mean_absolute_error(y_val, y_pred))
print("Baseline RMSE:", np.sqrt(mean_squared_error(y_val, y_pred)))
